src/backend/requirements.txt
# Core Framework
fastapi==0.110.3
gunicorn==22.0.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
greenlet==3.1.1

# Redis
redis[hiredis]==5.0.7

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0

# Task Queue
celery[redis]==5.4.0

# OTLP / Protobuf
opentelemetry-proto==1.24.0
protobuf==4.25.3
grpcio==1.64.1

# Encryption (for Fernet integration key storage)
cryptography==42.0.8

# Prometheus metrics
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
orjson==3.10.5
python-dateutil==2.9.0
uuid7==0.1.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.9.1
testcontainers[postgres,redis]==4.0.0rc1
aiosqlite==0.20.0
fakeredis==2.23.2
httpx-ws==0.6.0
